Remove useAsyncModal hook (#5)
Up to this point, this package provides a `useAsyncModal` hook that internally
calls `useAsyncModalState` and invokes a render func (provided as
parameter) while the modal is in the "open" state.

However, returning JSX from a hook is widely regarded as an
anti-pattern, and worse, it is presently unclear how to evolve the
`useAsyncModal` API going forward. As its main behavior can be easily
implemented as:

```tsx
<div>
  {state.isOpen && <MyComponent {...state.props} />
</div>
```

... this PR removes the `useAsyncModal` hook until a clear perspective
for such functionality emerges.